Refactor transformers
In `process_spec()` the transformers are applied to all keys within the spec (instead of only on the keys within the `data` dict). First `data_getter()` is run on the specification in `data`, then all `data` attributes are merged with the parent dict, finally transformers are run on all items in the parent dict.

In `make_werkvoorraad()` transformers are now passed in as a keyword argument (instead of hard-coding which transformers are used). Also `as_markdown_to_html()` was added as a transformer. It first converts a json stringva;ie to markdown format and then from there converts it to html.