Refactor `make_werkvoorraad()`
This commit changes `make_voorraad()` to take a `spec` object directly instead of a path. This is more flexible as a specification need not necessarily be loaded in from the file system but could also be created programmatically. `load_spec_from_json()` is still available to load the specification from a path.