Dispatch `wv-item-click` event when clicked
Whenever werkvoorraad-item receives a click it will dispatch an event containing the tag, id, classes and dataset of the element within the wc that was clicked.

The point of this is to be able to listen to this event and add additional functionality to the werkvoorraad. For example: open the query definition whenever the query info is clicked.